Adding a cleaned up matrix prep routine
Doesn't yet work. I believe it is because the inequality constraints need to be separated or else there are too few
and the program is unbounded. What needs to happen: for example for each node, you cannot count the number of
incoming nodes in total, you must count them using separate constraints. In the end the number of non-zero values
in the constraint matrix is the same, but there are more constraints.