The L best paths algorithm works
For some reason, the cvxopt internal solver doesn't work when only adjacent frames
are considered in determining if the number of nodes in and out is balanced. In this
case one must balance the number of incoming and outgoing nodes considering all
possible connections, so the constrain matrix is a bit less sparse.

If the glpk solver is used, only adjacent frames need be considered. It seems from
casual observation that this solver is slightly slower. Also, we are interested in
adding conical constraints for which this solver won't work.